Déplacement de la requête brouter dans un fichier
