+++ server.js
@@ -13,6 +13,7 @@ import fetchExceptions from './fetchExceptions'
 import { compute } from './geoStudio.js'
 import villes from './villesClassées'
 import { shuffleArray } from './utils'
+import brouterRequest from './brouterRequest'
 
 dotenv.config()
 
@@ -63,51 +64,9 @@ const cache = apicache.options({
 	debug: false,
 }).middleware
 
-const host = 'http://localhost:17777'
-//const host = 'https://brouter.de'
-
 app.get('/bikeRouter/:query', cache('1 day'), (req, res) => {
 	const { query } = req.params
-	const url = `${host}/brouter?lonlats=${query}&profile=safety&alternativeidx=0&format=geojson`
-	console.log('will fetch', query, url)
-
-	// For a reason I don't get, after 30 min of searching, using my local brouter server fails with the fetch function for a malformed header reason...
-	if (host.includes('localhost')) {
-		http
-			.get(url, (resp) => {
-				let data = ''
-
-				// Un morceau de réponse est reçu
-				resp.on('data', (chunk) => {
-					data += chunk
-				})
-
-				// La réponse complète à été reçue. On affiche le résultat.
-				resp.on('end', () => {
-					try {
-						res.json(JSON.parse(data))
-					} catch (e) {
-						console.log('error parsing locally', url)
-						console.log(e)
-					}
-				})
-			})
-			.on('error', (err) => {
-				console.log('Error: ' + err.message)
-			})
-	} else {
-		fetch(url)
-			.then((response) => {
-				console.log('did fetch from brouter', query)
-				const json = response.json()
-				console.log('json', json)
-				return json
-			})
-			.then((json) =>
-				// do some work... this will only occur once per 5 minutes
-				res.json(json)
-			)
-	}
+	brouterRequest(query, res.json)
 })
 
 const publicTransportRequestCore = `
